# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anki Card Generator
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  deckId: number;
  front: string;
  back: string;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}