# lib/api-zod/src/generated/types/createDeckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anki Card Generator
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDeckBody {
  name: string;
  /** @nullable */
  description?: string | null;
}