# lib/api-zod/src/generated/types/generateCardsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anki Card Generator
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from "./card";
import type { Deck } from "./deck";

export interface GenerateCardsResponse {
  deck: Deck;
  cards: Card[];
  generatedCount: number;
}